Okay, P2P Auth is Done.

+++ src/core/framework/network/portable/IMachineAddress.h
@@ -26,7 +26,7 @@ public:
     
     virtual IMachineAddress* createCopy() = 0;
     
-    virtual size_t getHash() const = 0;
+    virtual uint64_t getHash() const = 0;
     
     virtual std::string toString() const = 0;
 };
