+++ src/core/framework/network/steam/NGSteamAddress.cpp
@@ -32,7 +32,7 @@ IMachineAddress* NGSteamAddress::createCopy()
     return new NGSteamAddress(m_steamID);
 }
 
-size_t NGSteamAddress::getHash() const
+uint64_t NGSteamAddress::getHash() const
 {
     return m_steamID.ConvertToUint64();
 }
